Add and implement relationships trait
